Attempt at fixing plugin crash
Attempting to fix plugin crashing when a captain leaves during picking. Now when a player disconnects during picking, it just sends the plugin back to warm-up state. This isn't the original intended behavior to avoid re-voting for maps. However, an admin currently has to restart the plugin anyway so this way the plugin should avoid crashing with the same functional behavior.